requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Flask==2.0.0
ibm-cos-sdk==2.10.0
ibm-cos-sdk-core==2.10.0
ibm-cos-sdk-s3transfer==2.10.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
torch==1.8.1
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
